Update auth controller and routes for user service, fix frontend error with response
