+++ peerprep/frontend/src/app/signup/page.tsx
@@ -34,8 +34,8 @@ export default function SignUpPage() {
                 body: JSON.stringify({ email, username, password, confirmPassword }),
             });
 
+            
             const data = await response.json();
-
             if (response.ok) {
                 const { accessToken, id, username, email, isAdmin, createdAt } = data.data;
                 // store user info in session storage so it can be used in other components
@@ -45,8 +45,7 @@ export default function SignUpPage() {
             } else if (response.status === 400) {
                 setErrors(data.errors);
             } else {
-                const error = await response.json();
-                console.error(error.message);
+                console.error(data);
             }
         } catch (error) {
             console.error("Unexpected error during sign up:", error);
